# app/docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "GenAI Demo - DDD 電商平台 API",
    "description" : "基於領域驅動設計（DDD）的電商平台 API\n\n## 架構特色\n- 遵循 DDD 分層架構\n- 六邊形架構（Hexagonal Architecture）\n- CQRS 和事件驅動設計\n- 豐富的領域模型\n\n## 主要功能模組\n- 訂單管理：完整的訂單生命週期\n- 支付處理：多種支付方式和狀態管理\n- 庫存管理：即時庫存追蹤和預留\n- 產品管理：產品資訊和分類\n- 促銷活動：多樣化的促銷規則\n- 定價策略：動態定價和佣金計算\n- 客戶管理：會員等級和偏好設定\n\n## 錯誤處理\n所有 API 都遵循統一的錯誤回應格式，包含錯誤代碼、訊息、時間戳記和詳細資訊。\n",
    "contact" : {
      "name" : "GenAI Demo Team",
      "url" : "https://github.com/humank/genai-demo",
      "email" : "contact@genai-demo.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://test-api.genai-demo.com",
    "description" : "測試環境"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
    "/api/products/{productId}" : {
      "get" : {
        "tags" : [ "產品管理" ],
        "summary" : "獲取單個產品",
        "description" : "根據產品ID獲取產品詳細資訊，包括產品名稱、描述、價格、分類、庫存狀態等完整資料。",
        "operationId" : "getProduct",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "產品唯一識別碼",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "PROD-001"
        } ],
        "responses" : {
          "200" : {
            "description" : "成功獲取產品資訊",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "example" : {
                    "success" : true,
                    "data" : {
                      "id" : "PROD-001",
                      "name" : "產品名稱",
                      "description" : "產品描述",
                      "price" : {
                        "amount" : 100.0,
                        "currency" : "TWD"
                      },
                      "category" : "電子產品",
                      "inStock" : true,
                      "stockQuantity" : 50
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "產品不存在",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "產品管理" ],
        "summary" : "更新產品資訊",
        "description" : "更新指定產品的基本資訊，包括產品名稱、描述、價格、貨幣和分類。所有欄位都是可選的，只更新提供的欄位。",
        "operationId" : "updateProduct",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "產品唯一識別碼",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "PROD-001"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProductRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "422" : {
            "description" : "業務規則驗證失敗",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "產品更新成功",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "example" : {
                    "success" : true,
                    "data" : {
                      "id" : "PROD-001",
                      "name" : "產品名稱",
                      "description" : "產品描述",
                      "price" : {
                        "amount" : 100.0,
                        "currency" : "TWD"
                      },
                      "category" : "電子產品",
                      "inStock" : true,
                      "stockQuantity" : 50
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "請求參數無效",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "產品不存在",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "產品管理" ],
        "summary" : "刪除產品",
        "description" : "根據產品ID刪除指定產品。此操作不可逆，請謹慎使用。刪除前會檢查產品是否存在相關訂單。",
        "operationId" : "deleteProduct",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "產品唯一識別碼",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "PROD-001"
        } ],
        "responses" : {
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "產品刪除成功",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "example" : {
                    "success" : true,
                    "message" : "產品已成功刪除"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "產品不存在",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "產品無法刪除（存在相關訂單）",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products" : {
      "get" : {
        "tags" : [ "產品管理" ],
        "summary" : "獲取產品列表",
        "description" : "分頁查詢產品列表，支援自訂頁碼和每頁數量。回傳產品基本資訊、價格、庫存狀態等資料。",
        "operationId" : "getProducts",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "頁碼，從0開始",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          },
          "example" : 0
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "每頁數量，最大100",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          },
          "example" : 20
        } ],
        "responses" : {
          "400" : {
            "description" : "請求參數無效",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "成功獲取產品列表",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "example" : {
                    "success" : true,
                    "data" : {
                      "content" : [ {
                        "id" : "PROD-001",
                        "name" : "產品名稱",
                        "description" : "產品描述",
                        "price" : {
                          "amount" : 100.0,
                          "currency" : "TWD"
                        },
                        "category" : "電子產品",
                        "inStock" : true,
                        "stockQuantity" : 50
                      } ],
                      "totalElements" : 100,
                      "totalPages" : 5,
                      "size" : 20,
                      "number" : 0,
                      "first" : true,
                      "last" : false
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/pricing/rules/product/{productId}" : {
      "get" : {
        "tags" : [ "定價管理" ],
        "summary" : "獲取產品的定價規則",
        "description" : "查詢指定產品的所有有效定價規則，包含基礎價格、折扣策略、佣金設定和有效期間。結果按優先級和有效期間排序。",
        "operationId" : "getPricingRulesForProduct",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "產品的唯一識別碼",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "prod-123e4567-e89b-12d3-a456-426614174000"
        } ],
        "responses" : {
          "400" : {
            "description" : "請求參數無效，如產品ID格式錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "指定的產品不存在",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "成功獲取產品定價規則列表",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PricingRuleDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consumer/products/{productId}" : {
      "get" : {
        "tags" : [ "消費者商品" ],
        "summary" : "獲取商品詳情",
        "description" : "獲取指定商品的詳細信息",
        "operationId" : "getProductDetail",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "StandardErrorResponse" : {
        "required" : [ "code", "message", "path", "timestamp" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "錯誤代碼",
            "example" : "VALIDATION_ERROR"
          },
          "message" : {
            "type" : "string",
            "description" : "錯誤訊息",
            "example" : "請求參數驗證失敗"
          },
          "timestamp" : {
            "type" : "string",
            "description" : "錯誤發生時間",
            "format" : "date-time"
          },
          "path" : {
            "type" : "string",
            "description" : "請求路徑",
            "example" : "/api/orders"
          },
          "details" : {
            "type" : "array",
            "description" : "詳細錯誤資訊",
            "nullable" : true,
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            }
          },
          "traceId" : {
            "type" : "string",
            "description" : "追蹤ID，用於問題排查",
            "nullable" : true,
            "example" : "abc123-def456-ghi789"
          }
        },
        "description" : "標準錯誤回應"
      },
      "FieldError" : {
        "required" : [ "field", "message" ],
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "欄位名稱",
            "example" : "customerId"
          },
          "message" : {
            "type" : "string",
            "description" : "錯誤訊息",
            "example" : "客戶ID不能為空"
          },
          "rejectedValue" : {
            "type" : "object",
            "description" : "拒絕的值",
            "nullable" : true,
            "example" : null
          },
          "code" : {
            "type" : "string",
            "description" : "錯誤代碼",
            "nullable" : true,
            "example" : "NotBlank"
          }
        },
        "description" : "欄位錯誤詳情",
        "nullable" : true
      },
...
      "ProductDimensions" : {
        "type" : "object",
        "properties" : {
          "length" : {
            "minimum" : 0,
            "exclusiveMinimum" : false,
            "type" : "number",
            "description" : "長度（公分）",
            "format" : "double",
            "example" : 15.99
          },
          "width" : {
            "minimum" : 0,
            "exclusiveMinimum" : false,
            "type" : "number",
            "description" : "寬度（公分）",
            "format" : "double",
            "example" : 7.67
          },
          "height" : {
            "minimum" : 0,
            "exclusiveMinimum" : false,
            "type" : "number",
            "description" : "高度（公分）",
            "format" : "double",
            "example" : 0.825
          }
        },
        "description" : "產品尺寸資訊"
      },
...
      "ProductStatus" : {
        "type" : "string",
        "description" : "產品狀態",
        "example" : "ACTIVE",
        "enum" : [ "ACTIVE", "INACTIVE", "DISCONTINUED", "DRAFT", "OUT_OF_STOCK", "ACTIVE", "INACTIVE", "DISCONTINUED", "DRAFT", "OUT_OF_STOCK" ]
      },
      "UpdateProductRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string",
            "description" : "產品名稱",
            "example" : "iPhone 15 Pro"
          },
          "description" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string",
            "description" : "產品詳細描述，支援Markdown格式",
            "example" : "最新款iPhone，配備A17 Pro晶片，支援USB-C接口，具備專業級攝影功能"
          },
          "price" : {
            "minimum" : 0.01,
            "exclusiveMinimum" : false,
            "type" : "number",
            "description" : "產品價格（必須大於0）",
            "format" : "decimal",
            "example" : 35900.0
          },
          "currency" : {
            "pattern" : "^[A-Z]{3}$",
            "type" : "string",
            "description" : "貨幣代碼（ISO 4217標準）",
            "example" : "TWD",
            "enum" : [ "TWD", "USD", "EUR", "JPY", "CNY", "KRW", "HKD", "SGD" ]
          },
          "category" : {
            "type" : "string",
            "description" : "產品分類類型",
            "example" : "ELECTRONICS",
            "enum" : [ "ELECTRONICS", "FASHION", "HOME_LIVING", "SPORTS_FITNESS", "BEAUTY_CARE", "FOOD_BEVERAGE", "BOOKS_STATIONERY", "TOYS_GAMES", "AUTOMOTIVE", "OTHER" ]
          },
          "status" : {
            "$ref" : "#/components/schemas/ProductStatus"
          },
          "stockQuantity" : {
            "minimum" : 0,
            "type" : "integer",
            "description" : "庫存數量",
            "format" : "int32",
            "example" : 100
          },
          "weight" : {
            "minimum" : 0,
            "type" : "number",
            "description" : "產品重量（公克）",
            "format" : "double",
            "example" : 187.0
          },
          "dimensions" : {
            "$ref" : "#/components/schemas/ProductDimensions"
          },
          "sku" : {
            "maxLength" : 50,
            "minLength" : 0,
            "pattern" : "^[A-Z0-9\\-_]*$",
            "type" : "string",
            "description" : "產品SKU（庫存單位）",
            "example" : "IPHONE15PRO-128GB-BLACK"
          },
          "barcode" : {
            "pattern" : "^\\d{12,13}$",
            "type" : "string",
            "description" : "產品條碼（EAN-13或UPC-A格式）",
            "example" : "1234567890123"
          },
          "brand" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string",
            "description" : "產品品牌",
            "example" : "Apple"
          },
          "model" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string",
            "description" : "產品型號",
            "example" : "iPhone 15 Pro"
          },
          "tags" : {
            "maxItems" : 20,
            "minItems" : 0,
            "type" : "array",
            "description" : "產品標籤列表（用於搜尋和分類）",
            "example" : [ "smartphone", "premium", "5G" ],
            "items" : {
              "maxLength" : 30,
              "minLength" : 0,
              "type" : "string",
              "description" : "產品標籤列表（用於搜尋和分類）",
              "example" : "[\"smartphone\",\"premium\",\"5G\"]"
            }
          },
          "images" : {
            "maxItems" : 10,
            "minItems" : 0,
            "type" : "array",
            "description" : "產品圖片列表",
            "items" : {
              "$ref" : "#/components/schemas/ProductImageDto"
            }
          },
          "attributes" : {
            "maxItems" : 50,
            "minItems" : 0,
            "type" : "array",
            "description" : "產品屬性列表（如顏色、尺寸、材質等）",
            "items" : {
              "$ref" : "#/components/schemas/ProductAttributeDto"
            }
          },
          "warrantyMonths" : {
            "maximum" : 120,
            "minimum" : 0,
            "type" : "integer",
            "description" : "產品保固期（月）",
            "format" : "int32",
            "example" : 12
          },
          "manufacturer" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string",
            "description" : "產品製造商",
            "example" : "Apple Inc."
          },
          "countryOfOrigin" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string",
            "description" : "產品原產地",
            "example" : "中國"
          }
        },
        "description" : "產品更新請求資料",
        "example" : {
          "name" : "iPhone 15 Pro",
          "description" : "最新款iPhone，配備A17 Pro晶片，支援USB-C接口",
          "price" : 35900.0,
          "currency" : "TWD",
          "category" : "ELECTRONICS",
          "status" : "ACTIVE",
          "stockQuantity" : 100,
          "weight" : 187.0,
          "dimensions" : {
            "length" : 15.99,
            "width" : 7.67,
            "height" : 0.825
          },
          "sku" : "IPHONE15PRO-128GB-BLACK",
          "barcode" : "1234567890123",
          "brand" : "Apple",
          "model" : "iPhone 15 Pro",
          "tags" : [ "smartphone", "premium", "5G" ],
          "images" : [ {
            "id" : "img-001",
            "url" : "https://example.com/images/iphone15pro-front.jpg",
            "altText" : "iPhone 15 Pro 正面圖",
            "type" : "PRIMARY",
            "sortOrder" : 1
          } ],
          "attributes" : [ {
            "name" : "顏色",
            "value" : "太空黑",
            "type" : "COLOR",
            "displayOrder" : 1,
            "isKey" : true
          } ]
        }
      },
...
      "PricingRuleDto" : {
        "type" : "object",
        "properties" : {
          "priceId" : {
            "type" : "string"
          },
          "productId" : {
            "type" : "string"
          },
          "promotionId" : {
            "type" : "string"
          },
          "finalPrice" : {
            "type" : "number",
            "format" : "double"
          },
          "currency" : {
            "type" : "string"
          },
          "discountPercentage" : {
            "type" : "number",
            "format" : "double"
          },
          "discountAmount" : {
            "type" : "number",
            "format" : "double"
          },
          "effectiveFrom" : {
            "type" : "string",
            "format" : "date-time"
          },
          "effectiveTo" : {
            "type" : "string",
            "format" : "date-time"
          },
          "active" : {
            "type" : "boolean"
          },
          "productCategory" : {
            "type" : "string",
            "enum" : [ "ELECTRONICS", "FASHION", "GROCERIES", "HOME_APPLIANCES", "BEAUTY", "SPORTS", "BOOKS", "TOYS", "AUTOMOTIVE", "HEALTH", "GENERAL" ]
          },
          "normalCommissionRate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "eventCommissionRate" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "請在此處輸入 JWT Token，格式：Bearer {token}",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKey" : {
        "type" : "apiKey",
        "description" : "請在此處輸入 API Key",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}